Make FocusModeManager + Repo functional
This removes that convoluted shite I had for setting DND automatically
by bringing the responsibility of tracking whether we should be in DND
or not out of DND itself. I'm naming that FocusMode. Life's good.

Biggest change in the history of changes here. And probably one of the
best ones.